+++ R/TimeStratPetersenDiagErrorWHSteel_fit.R
@@ -112,7 +112,7 @@ TimeStratPetersenDiagErrorWHSteel_fit <-
 # The steelhead are nice because 100% of hatchery fish are adipose fin clipped and no wild fish are adipose fin clipped
 # The "diagonal entries" implies that no marked fish are recaptured outside the (time) stratum of release
 #
-    version <- '2024-05-09'
+    version <- as.character(packageDate("BTSPAS"))
     options(width=200)
 
 # Input parameters are