+++ R/TimeStratPetersenNonDiagErrorNP_fit.R
@@ -126,7 +126,7 @@ TimeStratPetersenNonDiagErrorNP_fit<- function( title="TSPNDENP", prefix="TSPNDE
   ## strata later. Transisions of marked fish are modelled non-parametrically.
   ##
   
-  version <- '2024-05-09'
+  version <- as.character(packageDate("BTSPAS"))
   options(width=200)
 
   ## Input parameters are